translation-service/requirements.txt
Flask==3.0.0
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
protobuf==4.25.1
redis==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
prometheus-client==0.19.0
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0


